Add Slot representation. An Inventory is a list of Slots.
