+++ src/main/java/com/cdl/inventorymanager/InventoryManagerApplication.java
@@ -5,9 +5,7 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
 public class InventoryManagerApplication {
-
 	public static void main(String[] args) {
 		SpringApplication.run(InventoryManagerApplication.class, args);
 	}
-
 }
