+++ src/main/java/com/cdl/inventorymanager/inventory/Inventory.java
@@ -1,69 +1,22 @@
 package com.cdl.inventorymanager.inventory;
 
-import java.util.Date;
+import java.util.List;
 
 public class Inventory {
-    private Long id;
-    private String name;
-    private Long quantity;
-    private String section;
-    private Date verificationDate;
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public Long getQuantity() {
-        return quantity;
-    }
-
-    public void setQuantity(Long quantity) {
-        this.quantity = quantity;
-    }
-
-    public String getSection() {
-        return section;
-    }
-
-    public void setSection(String section) {
-        this.section = section;
-    }
-
-    public Date getVerificationDate() {
-        return verificationDate;
-    }
-
-    public void setVerificationDate(Date verificationDate) {
-        this.verificationDate = verificationDate;
-    }
+    private List<Slot> slots;
 
     public Inventory() {
     }
 
-    public Inventory(String name, Long quantity, String section, Date verificationDate) {
-        this.name = name;
-        this.quantity = quantity;
-        this.section = section;
-        this.verificationDate = verificationDate;
+    public Inventory(List<Slot> slots) {
+        this.slots = slots;
     }
 
-    public Inventory(Long id, String name, Long quantity, String section, Date verificationDate) {
-        this.id = id;
-        this.name = name;
-        this.quantity = quantity;
-        this.section = section;
-        this.verificationDate = verificationDate;
+    public List<Slot> getSlots() {
+        return slots;
+    }
+
+    public void setSlots(List<Slot> slots) {
+        this.slots = slots;
     }
 }
