+++ src/main/java/com/cdl/inventorymanager/inventory/InventoryService.java
@@ -7,9 +7,10 @@ import org.springframework.stereotype.Service;
 
 @Service
 public class InventoryService {
-    public List<Inventory> getInventories() {
+    private final Inventory mockInventory = new Inventory(List.of(new Slot(1l, "Slot_1", 100l, "Section_1", new Date()),
+            new Slot(2l, "Slot_2", 200l, "Section_2", new Date())));
 
-        return List.of(new Inventory(1l, "Inventory1", 100l, "Section1", new Date()),
-                new Inventory(2l, "Inventory2", 200l, "Section2", new Date()));
+    public Inventory getInventory() {
+        return mockInventory;
     }
 }
